[types] - refactor: move DataSourceIntegration type to shared types
 - Extracted DataSourceIntegration type definition from front-end specific files into a shared types module
 - Removed redundant declaration of DataSourceIntegration in the front-end component file

[front] - refactor: clean up import of DataSourceIntegration type

 - Updated import statements to use the DataSourceIntegration type from the shared types module
 - Removed the local declaration of DataSourceIntegration to avoid duplication and maintain a single source of truth
