+++ front/pages/w/[wId]/builder/data-sources/managed.tsx
@@ -18,6 +18,7 @@ import {
 } from "@dust-tt/sparkle";
 import type {
   ConnectorProvider,
+  DataSourceIntegration,
   DataSourceType,
   EditedByUser,
   ManageDataSourcesLimitsType,
@@ -42,7 +43,6 @@ import { useRef } from "react";
 import { useContext, useEffect, useMemo, useState } from "react";
 import * as React from "react";
 
-import type { DataSourceIntegration } from "@app/components/data_source/DataSourceEdition";
 import ConnectorSyncingChip from "@app/components/data_source/DataSourceSyncChip";
 import { RequestDataSourcesModal } from "@app/components/data_source/RequestDataSourcesModal";
 import { subNavigationBuild } from "@app/components/navigation/config";
